DBLab3/app/requirements.txt
psycopg2-binary==2.9.5
schedule==1.2.0
Click==7.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
plotly==4.2.1
retrying==1.3.3
six==1.12.0
SQLAlchemy==1.3.10
Werkzeug==0.16.0
WTForms==2.2.1
gunicorn==19.9.0
redis==3.5.3
